requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.3
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
textblob==0.19.0
tqdm==4.67.1
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
